+++ docs/Whats-New.md
@@ -53,6 +53,7 @@ New:
 - Customizable OpenTopped properties (by Otamaa)
 - Automatic Passenger Deletion (by FS-21)
 - Script Action 74 to 81 for new AI attacks (by FS-21)
+- Script Actions 82 & 83 for modifying AI Trigger Current Weight (by FS-21)
 - Script Action to regroup temporarily around the Team Leader (by FS-21)
 
 Vanilla fixes: